refactor: move SodiumFakeBlockRenderer to compat.sodium.client package
