+++ src/main/java/yesman/epicfight/compat/iris/IRISCompat.java
@@ -5,7 +5,7 @@ import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import yesman.epicfight.client.events.engine.RenderEngine;
-import yesman.epicfight.client.renderer.SodiumFakeBlockRenderer;
+import yesman.epicfight.compat.sodium.client.SodiumFakeBlockRenderer;
 import yesman.epicfight.client.renderer.shader.compute.loader.ComputeShaderProvider;
 import yesman.epicfight.compat.ICompatModule;
