+++ src/main/java/yesman/epicfight/compat/sodium/client/SodiumFakeBlockRenderer.java
@@ -1,4 +1,4 @@
-package yesman.epicfight.client.renderer;
+package yesman.epicfight.compat.sodium.client;
 
 import java.nio.ByteBuffer;
 import java.nio.IntBuffer;
@@ -38,6 +38,8 @@ import net.minecraft.world.phys.Vec3;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.neoforge.client.model.data.ModelData;
+import yesman.epicfight.client.renderer.EpicFightRenderTypes;
+import yesman.epicfight.client.renderer.FakeBlockRenderer;
 
 @OnlyIn(Dist.CLIENT)
 public class SodiumFakeBlockRenderer implements FakeBlockRenderer {
